Separate ConfigTokenizer in the UML diagram
Now, there are ConfigCheckTokenizer and ConfigParseTokenizer.

The ConfigCheckTokenizer class is for checks for correctness of the
config file. It may involve specific parsing to tokens that has no
relation to parsing.

The ConfigParseTokenizer class is for parsing to get values from file
and this class has several methods implementing scanning
layers. Firstly, we scan the contents for separate blocks without
whitespace. Secondly, we scan resulting blocks and make tokens from
them. Thirdly, we scan resulting tokens for values.